requirements.txt
scikit-learn==1.3.0
spacy==3.7.2
numpy==1.24.3
streamlit==1.28.1
psycopg2-binary==2.9.10
requests==2.32.5
python-dotenv==1.1.1